ASP.NET Core MVC
A sandbox project written by following along with Microsoft Virtual
Academy's latest courses on the new ASP.NET Core MVC framework. This is
a basic CRUD web app which stores model objects in a local MSSQL
database, and contains functionality for authorization through Facebook.
The application is not meant for production. It's value lies in the code
written, to demonstrate where my current learning focus is.